Unify the request-field validation helpers into one module
Folds the per-period notionals guards (parseOptionalNotionals,
rejectUnsupportedNotionals) in beside the presence guards, so all reusable
request-field validation lives in one src/common/request_validation.h organized
into single-field, sub-table and vector sections. The narrowly-named
require_field.h and leg_notionals.h are removed; function names and behavior are
unchanged, so call sites only swap their include. The header documents what it
deliberately does not cover (string-identity fields and the price/rate/quote_id
either-or), which keep their own guards.

Full suite green (682 cases, unchanged).
