+++ src/mappers/bootstrap_curves_mapper.cpp
@@ -9,7 +9,7 @@
 #include "enum_convert.h"
 #include "error.h"
 #include "grid_utils.h"
-#include "require_field.h"
+#include "request_validation.h"
 
 namespace quantra {
