+++ src/mappers/floating_rate_bond_mapper.cpp
@@ -9,8 +9,7 @@
 #include "yield_parser.h"
 #include "enum_convert.h"
 #include "error.h"
-#include "leg_notionals.h"
-#include "require_field.h"
+#include "request_validation.h"
 
 namespace quantra {
