+++ src/mappers/ois_swap_mapper.cpp
@@ -5,8 +5,7 @@
 #include "schedule_parser.h"
 #include "enum_convert.h"
 #include "error.h"
-#include "leg_notionals.h"
-#include "require_field.h"
+#include "request_validation.h"
 
 namespace quantra {
