+++ src/mappers/swaption_mapper.cpp
@@ -11,8 +11,7 @@
 #include "curve_bootstrapper.h"
 #include "enum_convert.h"
 #include "error.h"
-#include "require_field.h"
-#include "leg_notionals.h"
+#include "request_validation.h"
 #include "eval_date_guard.h"
 #include "index_registry_builder.h"
 #include "swaption_generated.h"
