+++ src/mappers/year_on_year_inflation_cap_floor_mapper.cpp
@@ -4,7 +4,7 @@
 #include "enum_convert.h"
 #include "schedule_parser.h"
 #include "error.h"
-#include "require_field.h"
+#include "request_validation.h"
 
 namespace quantra {
