+++ src/mappers/year_on_year_inflation_swap_mapper.cpp
@@ -6,7 +6,7 @@
 #include "schedule_parser.h"
 #include "enum_convert.h"
 #include "error.h"
-#include "require_field.h"
+#include "request_validation.h"
 
 namespace quantra {
