+++ src/mappers/zero_coupon_inflation_swap_mapper.cpp
@@ -5,7 +5,7 @@
 #include "date_convert.h"
 #include "enum_convert.h"
 #include "error.h"
-#include "require_field.h"
+#include "request_validation.h"
 
 namespace quantra {
