+++ src/mappers/zero_coupon_swap_mapper.cpp
@@ -3,7 +3,7 @@
 #include "date_convert.h"
 #include "enum_convert.h"
 #include "error.h"
-#include "require_field.h"
+#include "request_validation.h"
 
 namespace quantra {
