+++ src/market/grid_utils.cpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 #include <sstream>
 #include "date_convert.h"
-#include "require_field.h"
+#include "request_validation.h"
 
 namespace quantra {
 namespace grid_utils {
