+++ src/market/pricing_registry.cpp
@@ -22,7 +22,7 @@
 
 #include "quote_registry.h"
 #include "inflation_curve_parsers.h"
-#include "require_field.h"
+#include "request_validation.h"
 
 #include <unordered_set>
