+++ src/market/swap_index_registry.cpp
@@ -1,6 +1,6 @@
 #include "swap_index_registry.h"
 
-#include "require_field.h"
+#include "request_validation.h"
 
 namespace quantra {
