+++ src/parsers/cms_leg_parser.cpp
@@ -3,7 +3,7 @@
 #include <cmath>
 #include <limits>
 
-#include "require_field.h"
+#include "request_validation.h"
 
 #include "vanilla_swap_generated.h"
