+++ src/parsers/term_structure_parser.cpp
@@ -9,7 +9,7 @@
 
 #include <cmath>
 
-#include "require_field.h"
+#include "request_validation.h"
 
 using namespace QuantLib;
