+++ src/parsers/term_structure_point_parser.cpp
@@ -6,7 +6,7 @@
 #include <ql/instruments/makeois.hpp>
 #include <ql/settings.hpp>
 
-#include "require_field.h"
+#include "request_validation.h"
 
 using namespace QuantLib;
